requirements.txt
twikit==1.7.6
pydantic-ai
google-generativeai
PyGithub
aiohttp
beautifulsoup4
pytz
python-dotenv
